frontend/pyproject.toml
[project]
name = "site-reporter-frontend"
version = "0.1.0"
description = "Streamlit frontend for the construction site report generator."
readme = "../README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.36.0",
    "requests>=2.32.0",
    "python-dotenv>=1.0.1",
    "pandas>=2.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

